pip-requirements.txt
Django == 3.2
Pillow == 8.2.0
django-admin-numeric-filter == 0.1.2
django-countries == 7.1
django-admin-rangefilter == 0.8.1
django-admin-list-filter-dropdown == 1.0.3
channels == 3.0.3
chess == 1.5.0
django-crispy-forms == 1.11.2
djangorestframework == 3.12.4
django-cors-headers == 3.7.0